chore: Reorganize some more files
