+++ tests/renderer/components/commands-publish-button-spec.tsx
